+++ packages/bigquery/index.cjs
@@ -1,5 +1,28 @@
 const { BigQuery } = require('@google-cloud/bigquery');
-const { EvidenceType, TypeFidelity } = require('@evidence-dev/db-commons');
+const { EvidenceType, TypeFidelity, getEnv } = require('@evidence-dev/db-commons');
+
+const envMap = {
+	projectId: [
+		{ key: 'EVIDENCE_BIGQUERY_PROJECT_ID', deprecated: false },
+		{ key: 'BIGQUERY_PROJECT_ID', deprecated: false },
+		{ key: 'project_id', deprecated: true },
+		{ key: 'PROJECT_ID', deprecated: true }
+	],
+	credentials: {
+		clientEmail: [
+			{ key: 'EVIDENCE_BIGQUERY_CLIENT_EMAIL', deprecated: false },
+			{ key: 'BIGQUERY_CLIENT_EMAIL', deprecated: false },
+			{ key: 'client_email', deprecated: true },
+			{ key: 'CLIENT_EMAIL', deprecated: true }
+		],
+		privateKey: [
+			{ key: 'EVIDENCE_BIGQUERY_PRIVATE_KEY', deprecated: false },
+			{ key: 'BIGQUERY_PRIVATE_KEY', deprecated: false },
+			{ key: 'private_key', deprecated: true },
+			{ key: 'PRIVATE_KEY', deprecated: true }
+		]
+	}
+};
 
 const standardizeResult = async (result) => {
 	var output = [];
@@ -43,20 +66,10 @@ const getCredentials = async (database) => {
 			return creds;
 		} else {
 			const creds = {
-				projectId:
-					process.env['BIGQUERY_PROJECT_ID'] ||
-					process.env['project_id'] ||
-					process.env['PROJECT_ID'],
+				projectId: getEnv(envMap, 'projectId'),
 				credentials: {
-					client_email:
-						process.env['BIGQUERY_CLIENT_EMAIL'] ||
-						process.env['client_email'] ||
-						process.env['CLIENT_EMAIL'],
-					private_key: (
-						process.env['BIGQUERY_PRIVATE_KEY'] ||
-						process.env['private_key'] ||
-						process.env['PRIVATE_KEY']
-					).replace(/\\n/g, '\n')
+					client_email: getEnv(envMap, 'credentials', 'clientEmail'),
+					private_key: getEnv(envMap, 'credentials', 'privateKey')?.trim()
 				}
 			};
 			return creds;