+++ packages/core-components/src/lib/unsorted/ui/Deployment/EvidenceDeploy.svelte
@@ -47,6 +47,7 @@
 
 	ul {
 		list-style: none;
+		padding-inline-start: 0px;
 	}
 
 	ul li::before {