+++ packages/core-components/src/lib/unsorted/ui/Formatting/CurrencyFormatGrid.svelte
@@ -52,7 +52,7 @@
 		-webkit-appearance: none;
 		-moz-appearance: none;
 		appearance: none;
-		padding: 0.75em;
+		padding: 0.35em;
 		width: calc(100% + 14px);
 		border: 1px solid var(--grey-200);
 		font-family: var(--ui-font-family);
@@ -60,6 +60,7 @@
 		margin: 0.5em 0 1.5em -7px;
 		transition: all 400ms;
 		cursor: pointer;
+		font-size: 14px;
 	}
 	select:hover {
 		border: 1px solid var(--grey-300);