+++ packages/ui/.svelte-kit/output/client/_app/immutable/chunks/1.2012992e.js
@@ -1,0 +1,1 @@
+import{default as t}from"../entry/error.svelte.16be1c96.js";export{t as component};