+++ packages/ui/.svelte-kit/output/server/nodes/0.js
@@ -1,0 +1,8 @@
+
+
+export const index = 0;
+export const component = async () => (await import('../entries/pages/_layout.svelte.js')).default;
+export const file = '_app/immutable/entry/_layout.svelte.9c263bf7.js';
+export const imports = ["_app/immutable/entry/_layout.svelte.9c263bf7.js","_app/immutable/chunks/index.e8978f25.js"];
+export const stylesheets = ["_app/immutable/assets/_layout.57383754.css"];
+export const fonts = [];