+++ packages/ui/.svelte-kit/output/server/nodes/2.js
@@ -1,0 +1,8 @@
+
+
+export const index = 2;
+export const component = async () => (await import('../entries/pages/_page.svelte.js')).default;
+export const file = '_app/immutable/entry/_page.svelte.9b650d10.js';
+export const imports = ["_app/immutable/entry/_page.svelte.9b650d10.js","_app/immutable/chunks/index.e8978f25.js"];
+export const stylesheets = [];
+export const fonts = [];